Remove sleep before triggering fatal error on SLPI
Prefer to maintain the SLPI in the state when the error was encountered,
rather than flushing queues to the host. This was originally intended to
ensure that logs reach logcat, but we can now see the file and line
number where the fatal error was hit, which wasn't the case when this
functionality was originally added.

Test: use TimerWorld to trigger chreAbort
Change-Id: I801d2ac6d51712c7556d81b5d8b2d05e8b435549