Remove crash on flush request timeout
Flush requests can get dropped when the system is too busy and CHRE
should avoid crashing the system in this case so the system can recover
on its own.

Unfortunately, this is slightly dangerous as a stale request can be
completed later and cause CHRE to treat a new request as complete, but
there isn't a way to solve this until CHRE can tie its requests to an
identifier.

Bug: 135558440
Test: Load onto device with short timeout and verify when
    a timeout happens, an event is sent to nanoapps.
Change-Id: I51fc3d08a8841ce4058feaa150c93e9432e566ec