Add clang-format preupload config
This config will ensure that clang-format is run on any changes being
uploaded to gerrit. This'll help keep changes in line with our style and
reduce cognitive load remembering all the various style guidelines.

Additionally, configs have been added in directories that have generated,
hand crafted, and external code to leave the formatting as-is since the
formatting should be preserved in those cases. Some code was moved to
new directories to support adding a .clang-format file so the files could
be ignored properly.

Test: ./run_sim.sh and ./load_chre.sh
Change-Id: I7854078c68b19bdd74e1c35cd2df69e417babf47