Fix formatting manually before running clang-format
clang-format isn't great at merging string literals that it thinks may
be broken up on purpose so it causes some weird formatting issues for a
few files in the directory. Additionally, clang-format also messes with
some hand-crafted parts of the codebase so this CL surrounds those parts
with "clang-format off/on".

Test: ./load_chre.sh
Change-Id: I731cf991c07c87696391f89f8e790886d60d2e5e