Fix message processing in power test nanoapp
The power test nanoapp wasn't verifying messages from the host using the
proper methods. Instead, copy the logic used elsewhere in the CHRE
framework to verify the message buffer.

Additionally, when sending a message to the host, the nanoapp didn't
properly finish the builder which is responsible for aligning the buffer
before it's sent elsewhere.

Bug: 122972449
Test: Load w/ latest power test client and verify a timer message can be
    processed and a valid message is returned

Change-Id: Ibe8f1ddf27820930cfc61e103dbbf5b447e6373d