Create common makefile for CHRE daemon
Now that the FastRPC headers are available in the system/chre directory,
the daemon's makefile can be open sourced as well.

Additionally, move the init.rc for the daemon into system/chre as well
and have it pulled into the device image only when the daemon is built.

Bug: 141892763
Test: compile

Merged-In: If88b7baea8d82bd9a20c9f8a5682cc970b359f8e
Change-Id: If88b7baea8d82bd9a20c9f8a5682cc970b359f8e