Fix compilation error
An earlier change modified the parameter type for messageSize to size_t,
but that caused a log statement to have the wrong specifier for
messageSize. Instead of changing the log statement, modify
sendMessageToNanoappFromHost to cast messageSize to a uint32_t before
sending to craftNanoappMessageFromHost since the latter method needs it
as a uint32_t anyway.

Test: ./load_chre.sh && ./run_sim.sh
Change-Id: Ic66b4d9880aaf390cbe4164800ee335b674bd921