Refactor sensor stack in the CHRE framework
With the addition of the sensor PAL, the sensor stack can now be
refactored to contain a PlatformSensorManager that manages all the
sensors given by the PAL (or some other code). This CL introduces that
manager while also updating the Sensor / PlatformSensor classes to
support the new interface. The SensorRequestManager will be updated in a
child CL.

Once completely refactored, the CHRE framework will go away from having
each sensor know how to manage itself and the PlatformSensorManager will
be responsible for managing all sensors which will just become wrappers
around the chreSensorInfo struct with some additional information.

To start, the refactor will target the Linux simulator and then the SEE
/ SMGR code will be refactored in a child CL.

Bug: 139693714
Test: ./run_sim.sh with the all the CLs in the same topic
Change-Id: Ic15aad6505c417e7ae9f0b87dd0397e0fb71302b