Add initDefaultSize method to DynamicVector
In order to emulate std::vector's constructor to initialize a vector
with N default-constructed elements, add a initDefaultSize method
that allows a user of dynamic vector to insert N default-constructed
elements without any copies.

This method is useful for storing elements that are neither movable nor
copyable, but still must be placed in a vector.

Adding a constructor is unfortunately not possible as it requires the
ability to signal when an allocation fails through an exception which
isn't possible on CHRE.

Bug: 139693714
Test: unit tests
Change-Id: I6694b0aa447d1ba337022f681c21908f91d50a1a