Enabling ranging CHQTS and fix a validation error
Enables the WiFi ranging CHQTS test while fixing a validation error
related to checking the ranging result timestamp.

Some WiFi chipset firmwares may perform ranging at the same time as they
perform a WiFi scan so the ranging result returned may have a timestamp
similar to the scan returned earlier in the test. Modify the test logic
to allow for this to occur.

Bug: 139692633
Test: Run CHQTS
Change-Id: Iaa55bdb5f00948d3b505e62cb9c68be4e4df9f7f