Fix debug dump for sensors
While resolving comments for the SEE refactor, the sensor name was
reverted back to the sensor vendor name. This isn't as useful for the
debug dumps which previously were printing the sensor type name. Revert
that logic back as well in this CL to keep what's printed in the debug
dump the same.

Before change:
Sensors:
 STMicro LSM6DSR: mode=2 int=18446744073709551615 lat=0 nappId=4
 STMicro LSM6DSR: mode=1 int=20000000 lat=3200000000 nappId=1

 Last 2 Sensor Requests:
  ts=235944215237083 nappId=1 sensType=STMicro LSM6DSR mode=ActiveContinuous int=20000000 lat=3200000000
  ts=235938791183333 nappId=4 sensType=STMicro LSM6DSR mode=ActiveOneShot

After change:
Sensors:
 Instant Motion: mode=2 int=18446744073709551615 lat=0 nappId=4
 Vendor Type 3: mode=1 int=20000000 lat=3200000000 nappId=1

 Last 2 Sensor Requests:
  ts=258886831803125 nappId=1 sensType=Vendor Type 3 mode=ActiveContinuous int=20000000 lat=3200000000
  ts=258886640616770 nappId=4 sensType=Instant Motion mode=ActiveOneShot

Bug: 139693714
Test: ./load_chre.sh
Change-Id: Ic7c4a2bff575d344ccc0c88d581007c30e375fbb