Changes to sensor validator to support on change and one shot
Refactor some code to allow on change and one shot sensors to be
supported in chre cross validator. The datapoints will still be saved in
an array but the array will be only one or two datapoints on both AP and
CHRE side.

Also, update proto file to have a flag isContinuous in SensorStartMessage
which causes the nano app to configure on change sensors to be
configured to default interval and latency. The isContinuous value will
also be stored in mCrossValidatorState and used to decide if sensor
samples with timestamps from before the test started are valid since
on-change sensors may send old cached values.

Bug: 146052784
Test: Ensure existing chqts continuous sensor tests pass

Change-Id: If20e7e0b4c8a00f7d3f3576956e18ea4887ac21c