Add proximity sensor support to chre cross validator
Also proximity is a special case since it is the only chreSensorByteData type
sensor that is being validated and its isNear property needs to be
converted to the float values used by the AP side framework before being
sent back from the nanoapp so it has its own handlers helper methods
in the nanoapp.

AP side code will compare values of 1 as far for CHRE datapoint and 0
for near and will check that in the case that its far that AP will not
be too close to , and vice versa that near will not go to close to max
value.

Change bug number for todo to break up Manager class to new bug that was
filed separately for this purpose.

Bug: 146052784
Bug: 154271551
Test: proximity chqts test passes
Change-Id: I94319e6dc0580c7e4086304753eedcd642d37a74