Update bias configuration logic
Today, bias configuration logic in SensorRequestManager is broken due to
assumptions made about how the platform implementation behaves as it
assumes that bias events are always enabled and that configureBiasEvents
has no effect.

Modify the logic such that bias event requests are appropriately tracked
and enabled / disabled at the core level to avoid pushing this
requirement to the platform.

Bug: 150144912
Test: Load onto device and verify bias updates continue to work
Change-Id: I51b6fa598111ae3bbbf82560f7ce3a8e44e4733d