Change timestamp assert logic for bias events in CHQTS
CHQTS currently asserts that all sensor data timestamps must be after
the test started. However, bias event timestamps are when the bias was
changed, not when the event was generated so they should be validated
differently. Modify the CHQTS logic to assert that timestamps are not
within the future, but don't assert any conditions beyond that.

Bug: 150144912
Test: Run CHQTS with latest bias configuration logic
Change-Id: I5af7e9507f9bd965135f9e3ac1b6efa990edaece