Increase the light sensor value allowed difference value
com.google.android.gts.locationcontext.LocationContextHostTest#testContextHubCrossValidationSensorLightTest
was failing on taimen and crosshatch and this increase in threshold will
account for the possible difference that will be seen in values of
light sensor. Specifically the difference of 1/16 for requantization
error for a reported resolution of 1 is used as the max allowed
difference here.

Bug: 159602515
Test: test passes
Change-Id: I87f7a2701389340284f48e6658d69b7d8022272c