Add MAKEFILE_LIST as a dep for compilation
MAKEFILE_LIST contains all the included makefiles used by the build
system. If any of these are modified, the CHRE build system should
recompile the CHRE library to ensure it's compiled with the latest
compilation options. Adding the MAKEFILE_LIST var as a dependency
of any compile rules ensures the build system is made aware of a diff
in the latest makefiles and will regenerate compiled files in that case.

Bug: 160244469
Test: Compile and verify that files are regenerate if any parameters in
various makefiles change.

Change-Id: Ie4cb0c54f641e14cd94538ee295f8eebb410816b