Avoid stalling out GNSS request processing
Consider the following sequence of events:

 * Nanoapp A requests to disable location session
 * Disable request is issued to platform, response pending
 * Nanoapp B requests to enable location session
 * Request is queued until nanoapp A's request finished processing
 * Location setting changes to 'disabled', mSettingChangePending set to
   true
 * Response to A's request is received (success), and
   handleLocationSettingChange() is processed as a no-op
   (setting state is disabled and mPlatformEnabled is false)

At this point in the previous code, GNSS would get stuck in the disabled
state, because all other requests would see Nanoapp B's request pending
in mStateTransitions, assume that we're waiting on an async callback
from the platform layer, and enqueue their request.

Avoid this by processing pending requests immediately after
handleLocationSettingChange() if that function call was a no-op (no
internal request is pending as a result).

Test: compile
Change-Id: Ib7a774f6b5789671655641fe1638b6dee14a30e0