Immediately update sampling status when it's received
Similar to host wake state and bias events, sampling status should be
updated as soon as it's received off the main event loop so that
nanoapps querying that state will see the actual latest state and not
some stale state.

This can cause issues where the nanoapp will asynchronously get old
events before receving the one they asynchronously queried. A true fix
for this is elaborated on and will be tracked in b/175023618.

Fixes: 174607102
Test: Run CHQTS

Change-Id: I2c274e96ba95c11377df62402a244e151c7200be