Support The Global Microphone Disable Setting in CHRE
Support disabling CHRE's access to microphone audio when it's
disabled by a user setting. A source suspended notification is
sent on access disabled, with the reverse happening on the
resumption of source availability.

Bug: 174691697
Test: Start device with the setting disabled, verify audio data
delivery via audio_world nanoapp, toggle the flag via the adb
command, verify data delivery stops and the src-suspended event
is received, toggle flag again to verify src-available event is
received and data delivery resumes. Repeat the test with the
initial condition reversed, and it works as expected.

Change-Id: Ifd4cbbbfe16677c83854f3b5a4a984e91c69eb3a