Simulate on logs sent to host flag state on empty buffer
There are a number of instances where after the callbacks in the event
loop manager call sendLogsToHost the HostLink::sendLogsMessageV2 method
will not be called and therefore the onLogsSentToHost() method will not
be called by the HostLink. Therefore call the onLogsSentToHost method
inside the sendLogsToHost method if sendLogMessageV2 is never called so
the state of the pending flush flags is reset properly.

Bug: 176430278
Bug: 146164384
Test: Observe logs consistently logged from CHRE
Change-Id: Iff69313361610f9a682ad206eb7845f09d6ecf60