Define method for annotating message with permissions
In v1.5, CHRE is requiring nanoapps to declare what data is contained in
the message they are sending so that it can enforce Android host
endpoints have been granted the required permissions to consume that
data. This new method will behave the same as the existing
chreSendMessageToHostEndpoint while allowing nanoapps to declare this
information.

Bug: 166846988
Test: compile
Change-Id: I05d5e6c984195e352c740c149bc7063afc6f5f41