User Settings Changed Notifications API Definition
Nanoapps need to be notified when there's a change in a
foreground user setting, which might cause CHRE to
suspend/enable certain data/services. The goal is to
implement APIs to enable nanoapps to register for
notifications on the settings they're interested in,
and receive events when changes to said settings occur,
so they can be prepared for the upcoming
suspension/resumption of the associated data/services.

Bug: 183026207
Test: N/A
Change-Id: I0cb14667653d31ad73e12a471cd814eb63fe0d3a