[Napp] Make wifi cross validation test less strict
For devices with first API level before S there are certain scenarios
where the number of CHRE scan results seen are less than the number of
AP results seen from an active scan request. Create a new message to the
nanoapp that is sent during the SETUP step that tells the nanoapp
whether or not to use a threshold that allows CHRE to have a certain
percentage less than AP.

Also change the test to do an O(n^2) lookup matching each of the CHRE
scan results with a result in the AP scan results. This means that the
test will no longer fail just because the AP results came in a different
order.

The test still does not check fields beyond comparing the string SSID
field and the bytes BSSID field as those fields are less predictable. A
TODO has been left to solve this in a future CL.

Bug: 181164508
Bug: 184653034
Bug: 185026344
Bug: 185188753
Test: Test passes on a pre S device. Fails on new device due to
legitimate bug.

Change-Id: Icf6b603c7acd91c78c3398a4172364dedbda0bd3