Support targeting events to specific groups of apps
Up until this point, events could only be targeted to a single nanoapp
or all nanoapps that are registered for a given event type. With this
CL, events can now be targeted to a subset of nanoapps registered for a
given event type. This can be used to allow events that cause different
power usages depending on the receipient nanoapp to be further targeted
to only a subset of nanoapps that need to actually consume that event
type at that moment. As a concrete example, this can be used to replace
big-image sensor types in SLPI and the same sensor type can be reused
for (e.g. accel), but under the hood, the sensor driver can split
requests for accel between big-image and micro-image and deliver accel
data events to the right subset of nanoapps.

Bug: 177324066
Test: Load CHRE and verify apps continue to work with noone using the
new targeting yet

Change-Id: I4ee3801ea04d0840f8208aa49e9a113d88fcf552