Increase minimum nanoapp message MTU
Raise the limit to 1000 bytes for CHRE implementations supporting v1.5+,
but permit nanoapps to compile against a lower value if needed for
compatibility with legacy implementations.

Also, remove the CHRE_MESSAGE_TO_HOST_MINIMUM_MAX_SIZE definition, to
prevent nanoapps from mistakenly referencing it. This is only meant to
constrain the value of CHRE_MESSAGE_TO_HOST_MAX_SIZE set on a given
platform, which can be accomplished without exposing a macro.

Fixes: 186788954
Test: compile
Change-Id: I05dc232807fd39c9d462c26f7f098469b4e7ab4c