Add CHRE version
Define a function getChreVersionString that is defined by each platform
which returns a string that identifies the device and build version that
this CHRE object was build from. The core init logic logs this version
string.

The shared platform implementation places the the version string in an
unstable_id section in memory.

Bug: 166270600
Test: Builds for all CHRE platforms

Change-Id: Ice737565a2a4b9df2b3b4a35676245f9d4d5dfcc