Add the WiFi NAN PAL API at the end of the structure.
To maintain ABI stability, we must only add new APIs to the end of
the PAL structure, since the PAL binary might not be recompiled,
and we can invoke invalid function pointers. Move the recently added
nanRanging API to the end to avoid this problem.

Bug: 209506015
Test: run `atest PtsChreTestCases` on an affected device/platform
and ensure all tests pass.

Change-Id: Ie3dafb9a52ebf28c3e19802e8dd36fd11e26c3ac