Refactor the CHRE daemon.
Since not all daemon variants will want to use flatbuffers as the
codec scheme for communicating with CHRE, the daemon base class
is now isolated from any flatbuffers dependencies, delegating such
functionality to a new base class called `FbsDaemonBase`, off of
which all daemon variants that want to use this codec shall now
derive from.

Bug: 225050694
Test: Build all 3 currently supported CHRE daemon variants, load
them on (an appropriate) device, ensure that the daemon loads and
operates as expected.

Change-Id: Ic556f868abeb6767ab1f4e71e8e3c0f7be4c052d