Fix a race condition in the GNSS sim tests
GnssSubscriptionWithSettingChange was failing when  executed via atest.

The root cause is a race condition where the PAL was not notified of the LOCATION setting change
before the location session was stop.

The CL solves the issue by explicitely waiting for the PAL to be notified of the setting change by
waiting for the change to be reflected in `chrePalGnssIsLocationEnabled()`.

Fixes: 229249554

Test: atest --host chre_simulation_tests
Change-Id: I7e7ddf0d97ae869bdbdfedeb43345a665efeb8d4