Fix CHPP NAN sub/cancel handling
For async work, CHPP uses a response to indicate that the service
accepted the response and a notification to deliver the data. Fix
sub/sub_cancel to abide by this scheme to prevent duplicate RX when
using NAN since previously a response was used to deliver the callback
data when sub/cancel was run.

Bug: 206614765
Test: Run CHPP instance locally
Change-Id: I9fc666626cbc28451a6631885a450ea6871664ed