Restore SocketCallbacks class from HAL impl
The SocketCallbacks abstraction is needed because SocketClient utilizes
a sp<> reference to the SocketClient::ICallbacks object provided to it
as input. If the abstraction isn't present, then there's no way to
ensure a strong pointer reference is also held by
HalChreSocketConnection and a crash can result if the HAL instance is
destroyed due to incorrect destruction ordering.

Bug: 230922800
Test: CHQTS
Change-Id: I93127cac4f6ceddd27adeb11aafb9abc0822b66a