Add API for queued load/unload test.
A minor refactoring has been done: Previously the callback object is
created via the registerLoadUnloadClient(). Now the callback is created
by the internal class ContextHubClientTestCallback. In this way we can
better encapsulate the atomic counters inside of the callback class and
make the test method for queued load/unload less dependant on the class
fields so that it doesn't conflit with other test methods.

Note that the way how we verify the correctness of the test is the same
as before.

Test: manually run impacted tests and pass
Bug: 151159967
Change-Id: I4409d283ee321f1fcf335980f9dac4d989ce7c3b