Add an extra millisecond delay to the EmbOS timer
Since the EmbOS timer is a millisecond tick based one, there is a
possibility that the timer can expire slightly ahead of schedule
(from CHRE's perspective) due to a clipping during the ms->ns
time factor conversion. Add an additional millisecond to the
requested delay to account for this.

Bug: 237819962
Test: Run timer world, verify that timers don't expire (a tiny
amount of time) before they are supposed to.

Change-Id: I5669827e4b444e09d20a6a699128f659f143673d