Graceful exit for StHalLpmaHandler
In previous implementation, we rely on mCurrentLpmaEnbled==mTargetLpmaEnabled to stuck in wait and detach the thread. This could cause issue when StHalLpmaHandler has been destructed but the working thread is not in wait condition, causing the detached thread accessing a destructed variable. Current implementation added terminate condition in the worker thread and wait for it to join before destructing the main object.

Test: Tested using chre_power_test_client to enable/disable audio data on flame couple of times and log looks normal. Also did pkill chre and no error message occurs.
Fixes: 241015054
Change-Id: Ibb737b58983f8f4b817222b4a78841282dafaf12