Clarify log info when unregistering client
Current code will log out an error when unregistering client even if
every thing is doing the expected behavior since it sends a message to
closed client to check it is properly unregistered. This is confusion
for developer when analyzing the log. This CL added a new log to
indicated that the following error log is intended.

Test: Presubmit
Bug: 242549170
Change-Id: I80865f49fcc1049853817ff7a18fdec6105e6706