Add a dedicated signal handler thread to the daemon
The current implementation of the CHRE daemon spawns threads to
perform various activities, and only handles incoming signals in
the socket server. This caused any received signals to terminate
the process. To circumvent this, the socket server was put in a
TEMP_FAILURE_RETRY loop which caused the daemon to never exit on
any signal. To address this, a dedicated signal handler thread is
spawned early in the daemon initialization, which sets flags for
a graceful shutdown of the daemon. A side-effect of this is that
any future thread/process spawned by the daemon must now use a
boolean (or the shutdownRequested flag) as its loop control variable,
and said flag must be appropriately set in the signal handler thread.

Fixes: 236169862
Test: Run 'adb shell pkill chre', verify that it works as expected.
Also run 'adb shell "pkill -{s} chre"' and observe that the daemon
only restarts for s in {2, 9, 15} (the signals we want to handle),
and not for any other signal received. Repeat test for all CHRE
daemon variants.

Change-Id: I62e5db53c76d67682fd3b690f23f341f0fe091d5