Enable build configuration for SystemHealthMonitor
This CL defined a new type to command for host to send to CHRE to
configure SystemHealthMonitor during runtime. This is needed since CHRE
has no clue which type of build (user/userdebug/eng) it is so will rely
on the host, which is build by Android, to send the configuration once
CHRE and host established connection.

Test: Build on user/userdebug/eng and flash all on oriole, verified that
user version does not set HealthMonitor to crash and both userdebug and
eng set it to crash
Fixes: 245569659

Change-Id: I5e045ae423d17235820c1501714dd673c3844a10