Revert^2  "Add a dedicated signal handler thread to the daemon"
This reverts commit 8d3cefff850cb4cb3dfee2fa880f11b02a265346.

To fix the crash happening when EINTR is received, a check was added to
socket_server.cc that retries if EINTR is encountered.

Fixes: 236169862
Test: Run 'adb shell pkill chre', verify that it works as expected.
Also run 'adb shell "pkill -{s} chre"' and observe that the daemon
only restarts for s in {2, 9, 15} (the signals we want to handle),
and not for any other signal received. Repeat test for all CHRE
daemon variants.

Change-Id: I767eac62d9a7061c32abad07a76a96145e5bf16e