Add RawStorage helper & apply std::launder
Now that we've migrated to C++17, technically std::launder is needed for
reinterpret_cast from std::aligned_storage to a contained type.
Compilers are unlikely to break this code, but it helps ensure we're
compliant with the language standard.

Add a helper class (RawStorage) to centralize this code across the
containers, and apply std::launder directly in places where RawStorage
isn't appropriate.

Fixes: 257135512
Test: ./run_tests.sh
Change-Id: I140a63216431954b5e1953033c206aac9ca57be0