BLE PAL test: add a gTest module for testing basic BLE PAL functions
Implement gTest for BLE PAL functions. It currently consists of 2 tests
- check capabilities
- filtered scan

To pass this test, it is required to have an external BLE device
advertising BLE beacons with service data for either the Google
eddystone or fastpair UUIDs.

It currently only run on Linux host with CHRE simulation, but not yet
setup to run on a phone. It also does not validate event data since
the Linux simulation does not return any meaningful data.

Bug: 249577259
Test: ./run_pal_impl_tests.sh and observe console output
Change-Id: I59330905c3c2140e46a82fe352aca525f1e4a24a