Move CHRE_PATCH_VERSION to unstable_id
CHRE_PATCH_VERSION is tied to Wall time. Because of this, the patch version can change the prebuilt binary even when the code is unchanged.

By storing this value in a constant, placed with the unstable_id
attribute, we can have the prebuilt picker correctly ignore this change
when looking for updates.

Bug: 258681566
Test: build_chre.sh -b, disassembled review of libchre_slip_skel.so
Change-Id: I8d04484d21ab13d80e5f8e081b429bc23bd353ad