Fix the audio concurrency test's processing of events
There is a race condition between receiving a message from
the host and receiving the events. This removes the test
step condition.

There is a problem when two events are received, one
that we expect for suspended audio and one when the audio
resumes. This handles receiving multiple events
correctly.

Bug: 269118726
Test: Presubmits
Change-Id: If9545377ef0b3b0fd4df2e4a576413c26f27bbfc