Revert "Suspend highly flakey test causing presubmit rejections."
This reverts commit 2e0d765d9f47c50c430db1909dcf77de62003228.

We need to relax the requirements for data to be presented
due to the more realistic usage of TaskManager in the linux
PAL. 1 ms is not enought time to process, especially in the
environment where this is being processed.

I recommend for future tests to be increased by an order of
magnitude as the timeout is a worst-case failure. On a device,
we still deliver events in a best-effort manner, and we cannot
control how much execution time we get.

This also fixes a bug where we were not releasing the data
that is received after the 3 instances that we want. Previously,
we just ignored this data without freeing it, and the
AddressSanitizer would fail and cause more flakes.

Bug: 270179898
Test: Run the test -> flake fixed!
Change-Id: Ie9e98d4533dfae800a6805af7dfc6f361a591eb2