Convert filters to big-endian format
For nanoapps compiled against a v1.8+ version of CHRE that are running
on an older version of CHRE, convert the service data UUID filters to
big endian format before forwarding them to CHRE and convert them to
their original format after.

Bug: 285207430
Test: With v1.7 version CHRE running, verify that ble_world using OTA
format for service data filters receives the expected results.

Change-Id: Ice129f1c818b3aa19aa07651117e2f18b7ab0a46